AVIF-info library and test
libavifinfo is a standalone library that can be used to extract the
width, height, bit depth and number of channels from an AVIF payload.
It is small on purpose and made of avifinfo.h and avifinfo.c.

Licensing is inspired from libaom (boilerplates, LICENSE and PATENTS).

Add CMakeLists.txt as the build system configuration file.

Add tests/avifinfo_test.cc and a 1x1 white pixel AVIF image compressed
with libavif for testing purposes.

Add README.md.

Change-Id: I26244b666854523fce3e001d2be9f19574a4715f