Streamed input API
Refactor the code to:
- sequentially read all boxes once into AvifInfoInternalFeatures,
- after "ipma" and "iref", check if AvifInfoInternalFeatures contain
  all necessary information to output AvifInfoFeatures.
This allows to fetch just a few bytes at a time from the stream,
in order, in a contiguous manner and never twice the same data.
The drawback is the ~250-byte long AvifInfoInternalFeatures struct
that is stored on the stack to avoid dynamic allocation. It can be
decreased or increased to adjust the variety of handled input AVIFs.

Change-Id: I98ccf4c55706f6c88d65008ccbcbefc782d2be87