Call AvifInfoIdentify() in AvifInfoGetFeatures()
Having to call AvifInfoIdentify() before AvifInfoGetFeatures() is
counter-intuitive. It was justified by the stream API, to avoid
reading the bytes of the 'ftyp' box twice. It is not necessary for the
raw pointer API. Matching the requirements of AvifInfoGetFeatures()
and AvifInfoGetFeaturesStream() is less important than an easy-to-use,
straightforward standalone single call to identify the file type and
get the AVIF features.

Change-Id: Ib726df5446318ceedbc5b0918e2c3e1b154252b0